# perf(compose): Always return a Promise without async.